Реализовано получение и отрисовка данных в списке
