+++ app/src/main/java/com/sssofi0101/foodstoreapp/data/retrofit/RetrofitImpl.kt
@@ -31,26 +31,9 @@ class RetrofitImpl : FoodDataSource {
         return retrofit
     }
 
-    override fun getMeals(category: String) : Meals {
-        var meals: Meals = Meals(listOf())
-        getService().getMenu(category).enqueue(object : Callback<Meals> {
+    override fun getMeals(category: String) : Call<Meals> {
+        return getService().getMenu(category)
 
-            override fun onResponse(call: Call<Meals>, response: Response<Meals>) {
-                if (response.body() != null) {
-                    meals = response.body()!!
-                    //здесь добавляем данные в адаптер для RecyclerView
-                }
-                else {
-                    throw Exception("Произошла ошибка при получении данных")
-                    }
-            }
-
-            override fun onFailure(call: Call<Meals>, t: Throwable) {
-                throw Exception("Ошибка доступа к серверу",t)
-            }
-
-        })
-        return meals
     }
 
 }
