+++ app/src/main/java/com/sssofi0101/foodstoreapp/domain/interfaces/FoodDataSource.kt
@@ -1,7 +1,8 @@
 package com.sssofi0101.foodstoreapp.domain.interfaces
 
 import com.sssofi0101.foodstoreapp.domain.models.Meals
+import retrofit2.Call
 
 interface FoodDataSource {
-    fun getMeals(category:String) : Meals
+    fun getMeals(category:String) : Call<Meals>
 }
