+++ app/src/main/java/com/sssofi0101/foodstoreapp/presentation/adapters/MealAdapter.kt
@@ -49,4 +49,5 @@ class MealAdapter (private val mealList:ArrayList<Meal>) : RecyclerView.Adapter<
         mealList.add(meal)
         notifyItemInserted(itemCount+1)
     }
+
 }
